CEDS-976 Route user to add-transport-containers page
If the user has selected that he is using any container, then the
next page being presented will the /add-transport-caontainers where
we present the add to list pattern and the user can enter multiple
container ids.

If the user does not have any container to enter then the journey will
proceed to the /total-numbers-of-items page.