CEDS-1047 Implement Standard/FullDec journey
We have added a ActionRefiner as an action composition enhancement in
order to allows us to add context and processing to the HTTP request,
by providing which jorney type the user is in.

This results in an extra DB call, which although it may seem unnecessary,
provides the ability in the future to resume the journey and knowing
which journey the user is in.

At this stage we have decided to implement in one one controller
(`RepresentativeDetailsPageController`) as a proof of concept and
because we already have one story ready to be played that affects that
same controller.

collaborated with @MaciejRewera, @PatrykRudnicki and @raghu1978