CEDS-1047 Remove supplementary/standard from the URI
In order to allow Standard and Supplementary user journeys we had to
remove the `supplementary` and `standard` from the URI, because
duplicating the routes files will make Play always map the controller to
the last URI found on the `app.routes` file.

Doing so meant that the unit tests had to be updated as well.