CEDS-1146 Map WCO-DEC Metadata based on a Mapping Strategy
There is never an elegant way to maintain backwards compatibility in
code, so this is the best attempt at it:

* Create a Java and Scala WCO-DEC Mapping classes

* Classes renamed to represent which model (Scala or Java) they are
mapping too. This is better than using "Old" and "New" naming
conventions.

* Having two mappers, for a Scala(old) and Java(new) models, allows us
to make use of the Strategy design pattern when time comes to marshall
the WCO-DEC declaration.

* Again the reason why going with a WCO-DEC Java model is due to ease
of serialization/deserialization.