CEDS-1152 Change to optional page when user doesn't fill anything
This commit covers making the page as optinal page with fixes for the tests.

Moreover I removed scalacOptions during compilation to give a possibility to use deprecated messages (now it's warning, not error)

Two helper methods for views are deprecated. We're trying to achieve using getElementById not Css - so it's first step.