CEDS-1143 Check the form valition and align the mapping
After a conversation with @raghu1978 we have realised that a lot of the
mapping logic checks are overly defensive as the production code will
never allow data to be in a given state.

For example on a form where a EORI and an address is being requested,
only one or them will be requested and none of them can be
`None/Null/Empty`.