CEDS-1213 Fix the new GoodsItem mapping strategy
This part of the code is a complete mess.
We now have a WCO-DEC model tangled all the way to the forms and a copy
of that that same model. We will need to discuss how we store the
information from the screens in our DB, as this is too complex.