CEDS-1612 Change Request[_] to Request[AnyContent]
This is required as we should be using AnyContent instead of _ and it
comes in handy as the FormAction class requires a Request of AnyContent
in order to inspect it's contents.

Another way would be to use bespoke code on the ModelCacheable, but
after speaking with Patryk it made sense to just change the signature on
the controllers.