CEDS-1955 Clean up the Warehouse model
As part of splitting the Warehouse Details page into three different
pages, we have tried to reuse the model. This became a problem for
several reasons:
 - Convoluted code on the update cache methods
 - Only one attribute belonged to the Warehouse type
 - Different approach from the rest of the controllers/views

With that in mind we decided to create one single model per page/view as
per our previous approach and keep the model clean.
This change also forced a code refactor on the BE service and on the
Acceptance tests suite.

Relates to : CEDS-1950, CEDS-1953, CEDS-1954