CEDS-1976 New GDS Design System layout unit tests
As part of the spike CEDS-1976 New component for H1s, Fieldsets and Legends,
 we have attempted to bring as much of the new GDS Design System approach
 https://design-system.service.gov.uk/.

After chatting with @bambuchaAdm about the spike he pointed me in the
direction of https://github.com/hmrc/play-frontend-govuk, which is an
HMRC project to migrate the GDS design system into a Scala Play
application.

The spike revolved around bringing components of the
https://github.com/hmrc/play-frontend-govuk library to try and get as
close as possible with the prototype and making the page fully accessible,
at least as accessible as the prototype.

The Exports Declaration Service choice page was the designated page
to perform this spike.