CEDS-1126 Associate Data Cells with appropriate headers
As per https://webaim.org/techniques/tables/data#headers:

The scope attribute tells the browser and screen reader that everything
within a column that is associated to the header with scope="col" in
that column, and that a cell with scope="row" is a header for all
cells in that row.

All <th> elements should generally always have a scope attribute.