CEDS-2180 Fix font configuration problem in QA/Dev
In the remote environments (development, QA and Staging) we are not able
to render the barcode due to a NullPointerException caused by the lack
of font configuration in our JRE-headless

```
java.lang.NullPointerException: null
	at sun.awt.FontConfiguration.getVersion(FontConfiguration.java:1264)
	at sun.awt.FontConfiguration.readFontConfigFile(FontConfiguration.java:219)
	at sun.awt.FontConfiguration.init(FontConfiguration.java:107)
	at sun.awt.X11FontManager.createFontConfiguration(X11FontManager.java:774)
```

We believe this is being caused by the fact that we were printing the
MRN as part of the barcode image. We hope that by removing the MRN from
the barcode image we will avoid this issue.

The line that fixes the issue is:
`code128Bean.setMsgPosition(HumanReadablePlacement.HRP_NONE)`

with @bambuchaAdm