CEDS-2250 Add 'I don't know' radio option for Transport Leaving Border
This new option is added for Clearance journey only.
Business rules demands this page to be optional for the declaration, but
radio buttons does not allow to un-check your answer. This new option
has been introduced to that and to allow users to proceed without
choosing any answer that would be interpreted as a valid answer.