BAU - Ignore floobits files
The Floobits for IntelliJ IDEA plugin https://floobits.com/help/plugins/intellij
generates some config files that we need to git ignore.

Floobits is a real-time collaborative editing tool for text editors and
IDEs. It allows for a seamless remote pair-programming experience.

The exports team tried this as part of this ticket and we will be
measuring the tool.