CEDS-2366 Display Office Of Exit summary section as previously
Regardless of the office of exit being inside or outside the UK, in the
summary page we should continue with the same layout and simply show
data captured in the same Office of Exit label.

The `Change` link should also link back to the /office-of-exit page, so
that the user has the option to change the office of exit location.