CEDS-2298 Display Time always on British time zone
DMS sends the times in UTC time, but our BE microservice converts them
to `LocalDateTime` which strips them of the Zone information.

Converting the type to `ZonedDateTime` allows us to store the date from
DMS as `UTC` or `Zulu time` and then on the FE we just need to inform
the view that we want to render that time on British Time Zone
`Europe/London`.

This ensures that we will see the time as `GMT/UTC` and `BST` depending
on the time of the year.

This holds true as long as DMS returns their dates in UTC.