CEDS-2441 Make Consignor Eori mandatory
This page must be mandatory.
Additionally I removed unnecessary controller spec cases plus added missing model spec.
I changed hasEori to be not longer optional. This field always must be provided.
The hasEori field is not a part of the backend model, so those changes are compatible with current backend.