CEDS-2678 Make Confirmation page panel resizable
The issue here was that for Supplementary declaration the green panel
was too narrow causing text inside to overflow.

The way this solution works is as follows:
Solution depends on JavaScript to check current font size used by the
browser. Based on this information a class describing column width for
the main part of the page is set. The default width was changed to
full-width to address non-JavaScript users.

Solution was inspired by: https://alistapart.com/article/fontresizing/