CEDS-2755 Add a Pointer Based URL logic
In some instances the WCO Declaration pointer will be returned but it
can map to two different pages on our user journey, which is depending on the
declaration type and the questions answered.

This means that we need to dynamically send the user to the correct page
based on the above.

This is an approach that keep the responsibility of this task/decision
in the RejectionReason class.