Update content and link on rejected declarations page list
The story raised the issue of screen readers not reading the context of the content correctly. The knock on effect of this, was that content was injected into the change link with hidden text, along with the html markup for ordered list.

The fix was to replcae the ordered list with plain English to improve context and readability for all users.