CEDS-2670 Parse tariff codes and transform them for CDS Exports
This is an automated way of parsing the tariff codes and
transforming them into our own CDS Expors service.

Going forward we would need to ensure the URLs are the most recent ones
and update them on the `urls.js` file and then execute start.js in order
to get the most up to date codes.

This is simply an quick approach to produce quick wins.