[CEDS-2870] Fix New Tariff Link Issues
All tariff links in expanders are now shown on a new line.

Links on C21 journey for these pages are back to pointing
to the tariff guide homepage:
* /consignment-references
* /are-you-completing-this-declaration-on-behalf-of-another-agent
* /declaration-items-list

Removed tariff expander from these pages entirely:
* /declaration/container
* /declaration/declaration-items-list
* /declaration/items/e8e46038/packages-list
* /declaration/items/14b127ae/additional-taric-codes-list
* /declaration/items/14b127ae/national-additional-codes-list
* /declaration/containers/234234/seals

One exception to this rule is the '/declaration/declaration-items-list'
page that Liz has said this is an exception to the rule of
removing them everywhere for these types of pages.