[CEDS-2879] Make address validation more permissive
Before the address form validation restricted values
to be alphanumeric and latin characters only.

This is in contrast to the DMS Schema rules that allow
any characters and only limit on the field lengths
allowed.

These changes relax the implemented validation rules
to become inline with the DMS Schema rules.