BAU Fix caching bug
The bug was appearing when user viewed their
submitted Declaration and then moved back and
started a new Declaration. The viewed one was
still in the cache and they were actually
trying to change a Declaration that was already
submitted. Since this is not possible, user
would see "There is a problem with service" page.

The bug was present because of the way the data
for declaration viewing page was taken from cache.
Declaration ID was added to session and based on
that cache data was retrieved. But this ID was
never cleared from session and therefore users
could get stuck, not being able to create a new
declaration.

This fix changes the way cache is being retrieved
by passing the declaration ID in URL parameter
instead of session. The difference to the user
is that this declaration id is visible in the
browser. But this should not be a massive problem
since we have this id present in some other pages.